requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
SQLAlchemy==1.4.31
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
whitenoise==5.3.0
